chore: address comments, refactorx1
